requirements.txt
# TrendScout AI 2.0 - Python Dependencies

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.43.0
feedparser==6.0.11

# Data Processing
pandas==2.2.1
numpy==1.26.4
python-dateutil==2.9.0
tqdm==4.66.2
python-dotenv==1.0.1

# NLP & Classical IR
spacy==3.7.4
scikit-learn==1.4.1
rank-bm25==0.2.2
gensim==4.3.2
nltk==3.8.1

# Deduplication
datasketch==1.6.4

# Storage
jsonlines==4.0.0

# UI
streamlit==1.33.0

# Testing
pytest==8.1.1
pytest-cov==5.0.0

# Phase 2+ - uncomment when ready
# neo4j==5.20.0
# sentence-transformers==2.7.0
# faiss-cpu==1.8.0
# chromadb==0.5.0
# langchain==0.1.20
# openai==1.25.0
# anthropic==0.25.0
